:zap: int型主键改为long，全面采用snowflake算法

+++ blade-auth/src/main/java/org/springblade/auth/service/BladeUserDetails.java
@@ -33,7 +33,7 @@ public class BladeUserDetails extends User {
 	/**
 	 * 用户id
 	 */
-	private Integer userId;
+	private Long userId;
 	/**
 	 * 租户编号
 	 */
@@ -59,7 +59,7 @@ public class BladeUserDetails extends User {
 	 */
 	private String avatar;
 
-	BladeUserDetails(Integer userId, String tenantCode, String name, String roleId, String roleName, String avatar, String username, String password, boolean enabled, boolean accountNonExpired, boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities) {
+	BladeUserDetails(Long userId, String tenantCode, String name, String roleId, String roleName, String avatar, String username, String password, boolean enabled, boolean accountNonExpired, boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities) {
 		super(username, password, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities);
 		this.userId = userId;
 		this.tenantCode = tenantCode;
